Renamed IpTrafficClass to IpNumber and added u8 constants in ip_number.
This change is mainly to avoid confusion as there is a field
called `traffic_class` in IPv6 headers that has nothing to with the
values in this enum.

Even I got confused, so to avoid this in the future the enum
was renamed from `IpTrafficClass` to `IpNumber` which is closer
to the actual used name "Assigned Internet Protocol Numbers".